[MeshWriter3mf] Refactor to inherit from Writer3mfBase; add thumbnail support and improve metadata handling
